# docs/swagger.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: TerraformGUI
  description: A convenient way to use and learn Terraform
paths:
...
  /projects/{project_id}/variables:
    get:
      description: Returns a project's variables
      parameters:
        - in: path
          name: project_id
          schema:
            type: integer
          required: true
          description: the id of the current project
      responses:
        '200':
          description: Successfully returned the project's variables
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - project_id
                    - variable_id
                    - name
                  properties:
                    project_id:
                      type: integer
                    variable_id:
                      type: integer
                    name:
                      type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
    post:
      description: Creates a variable
      parameters:
        - in: path
          name: project_id
          schema:
            type: integer
          required: true
          description: the id of the current project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object 
              required:
                - name
                - data
              properties:
                name:
                  type: string
                data: 
                  type: object
      responses:
        '200':
          description: Successfully created a new variable
          content:
            application/json:
              schema:
                type: object 
                required:
                  - variable_id
                  - project_id
                  - name
                properties:
                  variable_id: 
                    type: string
                  project_id: 
                    type: integer
                  name:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
  /projects/{project_id}/variables/{variable_id}:
    put:
      description: Updates a variable
      parameters:
        - in: path
          name: project_id
          schema:
            type: integer
          required: true
          description: the id of the current project
        - in: path
          name: variable_id
          schema:
            type: integer
          required: true
          description: the id of the variable to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object 
              required:
                - name
                - data
              properties:
                name:
                  type: string
                data: 
                  type: object
      responses:
        '200':
          description: Successfully updated the required variable
          content:
            application/json:
              schema:
                type: object 
                required:
                  - variable_id
                  - project_id
                  - name
                properties:
                  variable_id:
                    type: integer
                  project_id: 
                    type: integer
                  name:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string

    delete:
      description: Deletes a variable
      parameters:
        - in: path
          name: project_id
          schema:
            type: integer
          required: true
          description: the id of the current project
        - in: path
          name: variable_id
          schema:
            type: integer
          required: true
          description: the id of the variable to delete
      responses:
        '200':
          description: Successfully deleted the required variable
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/lucaslehot/TerraformGUI/0.1
...